+++ simply-synonyms/src/background/contextMenus.js
@@ -18,6 +18,30 @@ export default function () {
     parentId: 'menuParent',
     contexts: ['page']
   })
+  browser.contextMenus.create({
+    title: 'Help',
+    id: 'helpParent',
+    parentId: 'menuParent',
+    contexts: ['page']
+  })
+  browser.contextMenus.create({
+    title: 'Help',
+    id: 'help',
+    parentId: 'helpParent',
+    contexts: ['page']
+  })
+  browser.contextMenus.create({
+    title: 'Settings and Options',
+    id: 'settings',
+    parentId: 'helpParent',
+    contexts: ['page']
+  })
+  browser.contextMenus.create({
+    title: 'Contact Us',
+    id: 'contact',
+    parentId: 'helpParent',
+    contexts: ['page']
+  })
 
   browser.contextMenus.onClicked.addListener((info, tab) => {
     switch (info.menuItemId) {
@@ -25,6 +49,16 @@ export default function () {
         if (tab.url === 'chrome://newtab/') {
           browser.tabs.update(tab.id, { url: `https://www.merriam-webster.com/` })
         } else browser.tabs.sendMessage(tab.id, { action: "openQuickSearch" })
+        break
+      case 'help':
+        browser.tabs.create({ url: browser.runtime.getURL('page/help.html')})
+        break
+      case 'settings':
+        browser.tabs.create({ url: browser.runtime.getURL('page/settings.html')})
+        break
+      case 'contact':
+        browser.tabs.create({ url: 'https://synonyms.bweb.app/contact' })
+        break
     }
   })
 }