+++ simply-synonyms/src/pages/extensionPage.js
@@ -1,0 +1,4 @@
+import './extensionPage.scss'
+import browser from 'browserApi'
+
+document.getElementById('version-text').innerText = `V${browser.runtime.getManifest().version}`